backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0
python-dotenv==1.0.0

# Vector database
chromadb==0.4.22

# Embeddings
sentence-transformers==2.2.2
torch==2.1.0

# LLM integration
requests==2.31.0

# Validation
jsonschema==4.20.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
loguru==0.7.2

# Authentication
pyjwt==2.8.1
email-validator==2.1.0
